Modified   polowijo.gosari.BusinessObject/Dto/MasterPetugasDto.cs Modified   polowijo.gosari.BusinessObject/polowijo.gosari.BusinessObject.csproj Modified   polowijo.gosari.Core/Enums.cs Modified   polowijo.gosari.DAL/DBEntities.Context.cs Modified   polowijo.gosari.DAL/DBEntities.edmx Modified   polowijo.gosari.DAL/DBEntities.edmx.diagram Modified   polowijo.gosari.DAL/IServices/IPekerjaServices.cs Modified   polowijo.gosari.DAL/Services/PekerjaServices.cs Modified   polowijo.gosari.DAL/master_petugas.cs Modified   polowijo.gosari.DAL/polowijo.gosari.DAL.csproj Modified   polowijo.gosari.bongkar.muat/MainWindow.xaml.cs Modified   polowijo.gosari.bongkar.muat/Mapper/CustomMapper.cs Modified   polowijo.gosari.bongkar.muat/Model/Master/PekerjaModel.cs Modified   polowijo.gosari.bongkar.muat/UI/HOME/Home_View.xaml Modified   polowijo.gosari.bongkar.muat/UI/HOME/Home_View.xaml.cs Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_MENU.xaml Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_MENU.xaml.cs Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_Details.xaml Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_Details.xaml.cs Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_Insert.xaml Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_Insert.xaml.cs Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_View.xaml Modified   polowijo.gosari.bongkar.muat/UI/MASTER/MASTER_PETUGAS/Petugas_View.xaml.cs Modified   polowijo.gosari.bongkar.muat/packages.config Modified   polowijo.gosari.bongkar.muat/polowijo.gosari.bongkar.muat.csproj

+++ polowijo.gosari.DAL/IServices/IPekerjaServices.cs
@@ -11,7 +11,7 @@ namespace polowijo.gosari.DAL.IServices
     {
         List<master_petugas> GetAll();
         void Save(MasterPetugasDto Dto);
-        master_petugas GetById(decimal Id);
-        void DeleteById(decimal Id);
+        master_petugas GetById(int Id);
+        void DeleteById(int Id);
     }
 }
